[ElementsDetail] Refactor data storage protocol to use encrypted JSON strings
Previously, the application stored a encrypted serializable object in the database using SealedObjects. However, this caused the application to crash when the object was modified. To fix this issue, the code got refactored to store an encrypted JSON string instead.

Please note that previous versions of the application are not compatible with this commit due to the change in data storage format. See v1.0.0-rc02