[Sync] Warning before authentication
Prior to this change, the application would request user authentication before issuing a security warning when exporting passwords in CSV format. With this update, the order of these actions are reversed, ensuring that users now receive the security warning first, followed by the request for authentication.
This change enhances the user experience and prioritizes security awareness.